+++ CHANGELOG.md
@@ -9,6 +9,7 @@ have notable changes.
 Changes since v2.31
 
 - User attributes can now be retrieved from ID token and user_info endpoint. (#3028)
+- User attributes are now formatted better (a boolean becomes a checkbox, etc.) (#3103)
 
 ## v2.31 "Harmajankatu" 2022-11-29